Reorganize code into a separate general utils folder, create a higher order component to wrap tree elements in draggable props
